fixing

+++ 1811061037_PhamDangMinhTriet_BigSchool/Controllers/API/FollowingsController.cs
@@ -5,14 +5,13 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
-using System.Net.Http;
 using System.Web.Http;
 
 namespace _1811061037_PhamDangMinhTriet_BigSchool.Controllers.API
 {
     public class FollowingsController : ApiController
     {
-        private ApplicationDbContext _dbContext { get; set; }
+        private readonly ApplicationDbContext _dbContext;
 
         public FollowingsController()
         {
@@ -40,14 +39,19 @@ namespace _1811061037_PhamDangMinhTriet_BigSchool.Controllers.API
 
         }
         [HttpDelete]
-        public IHttpActionResult UnFollow(string followerId)
+        public IHttpActionResult UnFollow(string id)
         {
             var userId = User.Identity.GetUserId();
 
-            var follow = _dbContext.Followings.Single(c => c.FolloweeId == followerId && c.FolloweeId == userId);
+            var follow = _dbContext.Followings
+                .SingleOrDefault(a => a.FollowerId == userId && a.FolloweeId == id);
+
+            if (follow == null)
+                return NotFound();
 
             _dbContext.Followings.Remove(follow);
             _dbContext.SaveChanges();
+
             return Ok();
         }
